requirements.txt
httpx>=0.27.0
rich>=13.7.0
dnspython>=2.6.0
ipwhois>=1.2.0
